Реалиозован переход из логина к карте по клику
