+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"D:\\1. WEB\\2. courses\\5 course. React\\loft-taxi\\src\\index.js":"1","D:\\1. WEB\\2. courses\\5 course. React\\loft-taxi\\src\\reportWebVitals.js":"2","D:\\1. WEB\\2. courses\\5 course. React\\loft-taxi\\src\\App.jsx":"3","D:\\1. WEB\\2. courses\\5 course. React\\loft-taxi\\src\\components\\Exit.jsx":"4","D:\\1. WEB\\2. courses\\5 course. React\\loft-taxi\\src\\components\\Entrance.jsx":"5"},{"size":500,"mtime":1607505472113,"results":"6","hashOfConfig":"7"},{"size":362,"mtime":1607505472113,"results":"8","hashOfConfig":"7"},{"size":1202,"mtime":1608399349282,"results":"9","hashOfConfig":"7"},{"size":93,"mtime":1608283222789,"results":"10","hashOfConfig":"7"},{"size":1212,"mtime":1608383755317,"results":"11","hashOfConfig":"7"},{"filePath":"12","messages":"13","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"zn1hwq",{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"D:\\1. WEB\\2. courses\\5 course. React\\loft-taxi\\src\\index.js",[],"D:\\1. WEB\\2. courses\\5 course. React\\loft-taxi\\src\\reportWebVitals.js",[],"D:\\1. WEB\\2. courses\\5 course. React\\loft-taxi\\src\\App.jsx",[],"D:\\1. WEB\\2. courses\\5 course. React\\loft-taxi\\src\\components\\Exit.jsx",[],"D:\\1. WEB\\2. courses\\5 course. React\\loft-taxi\\src\\components\\Entrance.jsx",[]]
+[{"D:\\1. WEB\\2. courses\\5 course. React\\loft-taxi\\src\\index.js":"1","D:\\1. WEB\\2. courses\\5 course. React\\loft-taxi\\src\\reportWebVitals.js":"2","D:\\1. WEB\\2. courses\\5 course. React\\loft-taxi\\src\\App.jsx":"3","D:\\1. WEB\\2. courses\\5 course. React\\loft-taxi\\src\\components\\registry.jsx":"4","D:\\1. WEB\\2. courses\\5 course. React\\loft-taxi\\src\\components\\Login.jsx":"5","D:\\1. WEB\\2. courses\\5 course. React\\loft-taxi\\src\\components\\Profile.jsx":"6","D:\\1. WEB\\2. courses\\5 course. React\\loft-taxi\\src\\components\\Map.jsx":"7"},{"size":500,"mtime":1608474179630,"results":"8","hashOfConfig":"9"},{"size":362,"mtime":1607505472113,"results":"10","hashOfConfig":"9"},{"size":1619,"mtime":1608481577360,"results":"11","hashOfConfig":"9"},{"size":1765,"mtime":1608473773115,"results":"12","hashOfConfig":"9"},{"size":1244,"mtime":1608477413191,"results":"13","hashOfConfig":"9"},{"size":99,"mtime":1608393060004,"results":"14","hashOfConfig":"9"},{"size":91,"mtime":1608283183132,"results":"15","hashOfConfig":"9"},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"zn1hwq",{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"20","messages":"21","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"28","messages":"29","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"D:\\1. WEB\\2. courses\\5 course. React\\loft-taxi\\src\\index.js",[],"D:\\1. WEB\\2. courses\\5 course. React\\loft-taxi\\src\\reportWebVitals.js",[],"D:\\1. WEB\\2. courses\\5 course. React\\loft-taxi\\src\\App.jsx",[],"D:\\1. WEB\\2. courses\\5 course. React\\loft-taxi\\src\\components\\registry.jsx",[],"D:\\1. WEB\\2. courses\\5 course. React\\loft-taxi\\src\\components\\Login.jsx",[],"D:\\1. WEB\\2. courses\\5 course. React\\loft-taxi\\src\\components\\Profile.jsx",[],"D:\\1. WEB\\2. courses\\5 course. React\\loft-taxi\\src\\components\\Map.jsx",[]]
