--- src/components/Exit.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-export const Exit = () => {
-  return (
-    <>Exit</>
-  );
-}
