--- src/components/Exit.spec.jsx
@@ -1,11 +1,0 @@
-import React from 'react';
-import {Exit} from './Exit';
-import {render} from '@testing-library/react'
-
-
-describe("Exit", () => {
-  it("renders correctly", () => {
-    const {container} = render(<Exit />)
-    expect(container.innerHTML).toMatch("Exit")
-  })
-})
