+++ src/components/Login.spec.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import  Login  from './Login';
+import { render } from '@testing-library/react'
+
+
+describe("Login", () => {
+  it("renders correctly", () => {
+    const { getByLabelText } = render(<Login />);
+    expect(getByLabelText('Имя пользователя:')).toHaveAttribute('name', 'email')
+    expect(getByLabelText('Пароль*')).toHaveAttribute('name', 'password')
+  });
+});
